Selectively apply preffered vector width 512.
-mprefer-vector-width actually slowed down some benchmarks, e.g.
compressed-scan autovec by almost 50%. Performance analysis there
concludes that it is just microarchitecture weirdness, possible due to
worse port usage on the CPU. To circumvent this problem, we just apply a
"use 512 here" hint to the methods where we know it improves
performance.

Note: This is not actually a "minimum", as the name suggests: When
compiling for hardware without 512-bit registers, it will use the
available register width.